requirements.txt
beautifulsoup4==4.12.2
PyYAML==6.0
selenium==4.10.0
